+++ src/components/fields/Field_SingleRelationship.vue
@@ -46,7 +46,7 @@
             <span class="isDeadIndicator" v-if="localInput.isDead">
               †
             </span>
-            <span :class="{'isDead': localInput.isDead}">
+            <span :class="{'isDead': (localInput.isDead && !hideDeadCrossThrough)}">
                 {{stripTags(localInput.label)}}
             </span>
           </span>
@@ -145,7 +145,7 @@
                 <span class="isDeadIndicator" v-if="opt.isDead">
                   †
                 </span>
-                <span :class="{'isDead': opt.isDead}" v-html="opt.label">
+                <span :class="{'isDead': (opt.isDead && !hideDeadCrossThrough)}" v-html="opt.label">
                 </span>
               </q-item-label>
               <q-item-label caption class="text-cultured" v-html="opt.hierarchicalPath"></q-item-label>